# API/openapi.yaml
openapi: 3.0.0
info:
    title: Supplier, Warehouse, and Good Management API
    version: 1.0.0
    description: API for managing suppliers, warehouses, and goods
servers:
    - url: /api/v1 # Replace with your actual base path
components:
    schemas:
        supplier:
            type: object
            properties:
                _id:
                    type: string
                name:
                    type: string
                description:
                    type: string
                rating:
                    type: string
        postSupplier:
            type: object
            properties:
                name:
                    type: string
                description:
                    type: string
                rating:
                    type: number
...
        error:
            type: object
            properties:
                code:
                    type: number
                title:
                    type: string
                details:
                    type: string
paths:
    /supplier:
        get:
            summary: List all suppliers
            tags:
                - Supplier
            responses:
                '200':
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/supplier'
                            example:
                                - _id: '6711549715d86595bf6713a4'
                                  name: 'UAB DEŠRELĖS'
                                  description: 'Desšrelių tiekėjas'
                                  rating: 2
                                - _id: '671167bfc618eac8edafca78'
                                  name: 'UAB KAUNO GRŪDAI'
                                  description: 'Grūdų tiekėjas'
                                  rating: 5
                '403':
                    description: Forbidden
                    content:
                        application/json:
                            schema:
                                type: object
                                items:
                                    $ref: '#/components/schemas/error'
                            example:
                                code: 403
                                title: Forbidden
                                details: Not allowed
                '401':
                    description: Unauthorized
                    content:
                        application/json:
                            schema:
                                type: object
                                items:
                                    $ref: '#/components/schemas/error'
                            example:
                                code: 401
                                title: Unauthorized
                                details: Not allowed
            security:
                - bearerAuth: []

        post:
            summary: Create a new supplier
            tags:
                - Supplier
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/supplier'
                        example:
                            name: 'TESTSupplier'
                            description: 'TESTSUPP'
                            rating: 0

            responses:
                '201':
                    description: Supplier created successfully
                '400':
                    description: Bad Request - Invalid data provided
                    content:
                        application/json:
                            schema:
                                type: object
                                items:
                                    $ref: '#/components/schemas/error'
                            example:
                                code: 422
                                title: Bad Request
                                details: Malformed Body
                '422':
                    description: Conflict - Supplier with the same name already exists
                    content:
                        application/json:
                            schema:
                                type: object
                                items:
                                    $ref: '#/components/schemas/error'
                            example:
                                code: 422
                                title: Unprocessable Entity
                                details: Supplier with this name already exists
                '403':
                    description: Forbidden
                    content:
                        application/json:
                            schema:
                                type: object
                                items:
                                    $ref: '#/components/schemas/error'
                            example:
                                code: 403
                                title: Forbidden
                                details: Not allowed
                '401':
                    description: Unauthorized
                    content:
                        application/json:
                            schema:
                                type: object
                                items:
                                    $ref: '#/components/schemas/error'
                            example:
                                code: 401
                                title: Unauthorized
                                details: Not allowed

            security:
                - bearerAuth: []
    /supplier/{supplierId}:
        get:
            summary: Get a supplier by ID
            tags:
                - Supplier
            parameters:
                - name: supplierId
                  in: path
                  required: true
                  schema:
                      type: string
                  example: 6711549715d86595bf6713a4
            responses:
                '200':
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/supplier'
                                type: object

                            example:
                                _id: '6711549715d86595bf6713a4'
                                name: 'UAB DEŠRELĖS'
                                description: 'Desšrelių tiekėjas'
                                rating: 2

                '400':
                    description: Bad Request - Invalid supplier ID provided
                    content:
                        application/json:
                            schema:
                                type: object
                                items:
                                    $ref: '#/components/schemas/error'
                            example:
                                code: 400
                                title: Bad Request
                                details: Supplier id was not provided or is not a valid object id
                '404':
                    description: Not Found - Supplier with the provided ID not found
                    content:
                        application/json:
                            schema:
                                type: object
                                items:
                                    $ref: '#/components/schemas/error'
                            example:
                                code: 404
                                title: Not Found
                                details: Supplier with this id was not found
                '403':
                    description: Forbidden
                    content:
                        application/json:
                            schema:
                                type: object
                                items:
                                    $ref: '#/components/schemas/error'
                            example:
                                code: 403
                                title: Forbidden
                                details: Not allowed
                '401':
                    description: Unauthorized
                    content:
                        application/json:
                            schema:
                                type: object
                                items:
                                    $ref: '#/components/schemas/error'
                            example:
                                code: 401
                                title: Unauthorized
                                details: Not allowed
            security:
                - bearerAuth: []

        put:
            summary: Update a supplier by ID
            tags:
                - Supplier
            parameters:
                - name: supplierId
                  in: path
                  required: true
                  schema:
                      type: string
                      example: 6711549715d86595bf6713a4
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/postSupplier'
                        examples:
                            All parameters:
                                value:
                                    name: 'New Name'
                                    description: 'New description'
                                    rating: 5
                                description: While providing all parameters
                            Partial parameters:
                                description: While providing few parameters
                                value:
                                    description: 'New description'

            responses:
                '200':
                    description: Supplier updated successfully
                '400':
                    description: Bad Request - Invalid data provided
                    content:
                        application/json:
                            schema:
                                type: object
                                items:
                                    $ref: '#/components/schemas/error'
                            example:
                                code: 400
                                title: Bad Request
                                details: Supplier id was not provided or is not a valid object id
                '404':
                    description: Not Found - Supplier with the provided ID not found
                    content:
                        application/json:
                            schema:
                                type: object
                                items:
                                    $ref: '#/components/schemas/error'
                            example:
                                code: 404
                                title: Not Found
                                details: Supplier with this id was not found
                '403':
                    description: Forbidden
                    content:
                        application/json:
                            schema:
                                type: object
                                items:
                                    $ref: '#/components/schemas/error'
                            example:
                                code: 403
                                title: Forbidden
                                details: Not allowed
                '401':
                    description: Unauthorized
                    content:
                        application/json:
                            schema:
                                type: object
                                items:
                                    $ref: '#/components/schemas/error'
                            example:
                                code: 401
                                title: Unauthorized
                                details: Not allowed
            security:
                - bearerAuth: []

        delete:
            summary: Delete a supplier by ID
            tags:
                - Supplier
            parameters:
                - name: supplierId
                  in: path
                  required: true
                  schema:
                      type: string
                  example: 6711549715d86595bf6713a4
            responses:
                '204':
                    description: Supplier deleted successfully
                '400':
                    description: Bad Request - Invalid supplier ID provided
                    content:
                        application/json:
                            schema:
                                type: object
                                items:
                                    $ref: '#/components/schemas/error'
                            example:
                                code: 400
                                title: Bad Request
                                details: Supplier id was not provided or is not a valid object id
                '404':
                    description: Not Found - Supplier with the provided ID not found
                    content:
                        application/json:
                            schema:
                                type: object
                                items:
                                    $ref: '#/components/schemas/error'
                            example:
                                code: 404
                                title: Not Found
                                details: Supplier with provided id was not found
                '403':
                    description: Forbidden
                    content:
                        application/json:
                            schema:
                                type: object
                                items:
                                    $ref: '#/components/schemas/error'
                            example:
                                code: 403
                                title: Forbidden
                                details: Not allowed
                '401':
                    description: Unauthorized
                    content:
                        application/json:
                            schema:
                                type: object
                                items:
                                    $ref: '#/components/schemas/error'
                            example:
                                code: 401
                                title: Unauthorized
                                details: Not allowed
            security:
                - bearerAuth: []
...